numcodecs_swizzle_reshape

Function swizzle_reshape

pub fn swizzle_reshape<T: Copy, S: Data<Elem = T>>(
    data: ArrayBase<S, IxDyn>,
    axes: &[AxisGroup],
) -> Result<Array<T, IxDyn>, SwizzleReshapeCodecError>

Swizzle and reshape the input data array with the new axes.

Errors

Errors with

SwizzleReshapeCodecError::InvalidAxisIndex if any axis is out of bounds

SwizzleReshapeCodecError::InvalidAxisPermutation if the axes permutation does not contain every axis index exactly once

SwizzleReshapeCodecError::MultipleRestAxes if the axes permutation contains more than one Rest-axes marker