numcodecs_swizzle_reshape

Function undo_swizzle_reshape_into

pub fn undo_swizzle_reshape_into<T: Copy>(
    encoded: ArrayView<'_, T, IxDyn>,
    decoded: ArrayViewMut<'_, T, IxDyn>,
    axes: &[AxisGroup],
) -> Result<(), SwizzleReshapeCodecError>

Reverts the swizzle and reshape of the encoded array with the axes and outputs it into the decoded array.

Errors

Errors with

SwizzleReshapeCodecError::InvalidAxisIndex if any axis is out of bounds

SwizzleReshapeCodecError::InvalidAxisPermutation if the axes permutation does not contain every axis index exactly once

SwizzleReshapeCodecError::MultipleRestAxes if the axes permutation contains more than one Rest-axes marker

SwizzleReshapeCodecError::MismatchedDecodeIntoArray if the encoded array’s shape does not match the shape that swizzling and reshaping an array of the decoded array’s shape would have produced